Allow HTMLOutputStream to write to a `String`
Currently it's not possible to write the rendered `HTML` to a simple `String` instead of `stdout` or a file, even though `String` conforms to `TextOutputStream`.

Since `String` is a value type, not a reference type, passing it as a first argument makes a copy of that string and the stream writes to a copy. It would be possible to receive a copy that has the rendered content if `output` was a publicly readable property.